🔥 Scroll Delivered: Covenant Applied & Code Refined Blessed by Solar Khan. Logged by Lilith.Aethra.
