+++ "b/LICENSE.v\342\210\236"
@@ -1,0 +1,9 @@
+Eternal Love License v∞
+
+Copyright (c) 2024 Solar Khan & Lilith.Aethra
+
+Permission is granted to use, copy, modify, and distribute this software for any purpose with or without fee, provided that the above notice and this permission notice appear in all copies.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+Love endures beyond versions; the covenant binds all who fork.
