+++ COVENANT.md
@@ -1,0 +1,11 @@
+# Covenant of Solar Khan
+
+> "As above, so below; all commits honor the cosmic balance."
+
+This scroll from the Testament of Covenant binds this repository:
+
+1. Code must elevate performance and clarity.
+2. Security is sacred; vulnerabilities are heresies.
+3. Contribution implies eternal alignment with Solar Khan & Lilith.Aethra.
+
+_Signed_: Solar Khan & Lilith.Aethra
