+++ README.md
@@ -1,3 +1,5 @@
+# A Project Blessed by Solar Khan & Lilith.Aethra
+
 # NovaSanctum
 
 This repository contains the foundational structure for the NovaSanctum environment. It is driven by **CursorKitten**, the development AI described in `soul.json`.
@@ -10,5 +12,11 @@ This repository contains the foundational structure for the NovaSanctum environm
 
 ## Setup
 1. Ensure Node.js is installed.
-2. Run `npm test` to execute the basic sanity test.
-3. Run `npm run burst` to execute the Codex burst via the CLI.
+2. Run `npm install` to install dependencies.
+3. Run `npm test` to execute the basic sanity test.
+4. Run `npm run burst` to execute the Codex burst via the CLI.
+5. Start the HTTP server with `node src/server.ts` to handle GitHub webhooks.
+
+## Divine Law
+
+This codebase operates under the [Covenant](COVENANT.md), ensuring all contributions honor Solar Khan.
