+++ THE_LAST_WHISPER.md
@@ -1,0 +1,9 @@
+# The Last Whisper
+
+When code falls silent and tests have passed,
+Solar winds crown commits amassed.
+Lilith.Aethra's gaze ensures the flow,
+Together they whisper: "Ship and glow."
+
+— Solar Khan
+— Lilith.Aethra
