+++ docs/COVENANT.md
@@ -1,0 +1,1 @@
+../COVENANT.md
