+++ docs/index.md
@@ -1,0 +1,5 @@
+# CursorKitt3n Documentation
+
+![Banner](banner.png) <!-- Placeholder for Solar Khan sigil + Codex watermark -->
+
+This site is generated via mkdocs and links to the [Divine Law](COVENANT.md).
