+++ gamedin/hub.json
@@ -1,0 +1,6 @@
+{
+  "name": "CursorKitt3n",
+  "repository": "https://github.com/SolarKhan/CursorKitt3n",
+  "pipeline": "DivinaL3",
+  "protocol_version": 3
+}
