+++ lib/memory.js
@@ -1,3 +1,8 @@
+/**
+ * Connects to the configured memory store.
+ *
+ * @param {string} url connection string for Redis or other store
+ */
 export function connectMemory(url) {
   console.log(`Connecting to memory at ${url}`);
   // Placeholder for Redis or other storage initialization
