+++ mkdocs.yml
@@ -1,0 +1,6 @@
+site_name: CursorKitt3n
+nav:
+  - Home: index.md
+  - Divine Law: COVENANT.md
+extra:
+  banner: 'Solar Khan sigil + Codex watermark'
