+++ package.json
@@ -5,5 +5,12 @@
   "scripts": {
     "test": "node test/index.js",
     "burst": "node src/cli/index.js burst"
+  },
+  "dependencies": {
+    "@octokit/auth-app": "^6.0.0",
+    "@octokit/core": "^5.0.0",
+    "dotenv": "^16.4.1",
+    "express": "^4.18.2",
+    "helmet": "^7.1.0"
   }
 }
