+++ src/modules/EidolonCore/index.js
@@ -1,3 +1,8 @@
+/**
+ * Initializes the EidolonCore subsystem.
+ *
+ * Used to bootstrap shared engine state across GameDin modules.
+ */
 export function initEidolonCore() {
   console.log('EidolonCore initialized');
 }
