+++ src/modules/GameDin/index.js
@@ -1,3 +1,6 @@
+/**
+ * Sets up GameDin-specific frontend logic bindings.
+ */
 export function initGameDin() {
   console.log('GameDin frontend logic loaded');
 }
