+++ src/modules/Sovereign/index.js
@@ -1,3 +1,8 @@
+/**
+ * Bootstraps the Sovereign orchestration layer.
+ *
+ * Intended for high-level automation flows.
+ */
 export function initSovereign() {
   console.log('Sovereign orchestration active');
 }
